% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{append.slash}
\alias{append.slash}
\title{function adds to trailing path separator at the end of dir.path}
\usage{
append.slash(dir.path)
}
\arguments{
\item{dir.path}{string any path with or without trailing slash}
}
\value{
dir.path string same path with slash at the end
}
\description{
function adds to trailing path separator at the end of dir.path
}

