% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute.metrixes}
\alias{compute.metrixes}
\title{Function compute.metrixes given confusion matrix, computes precision, accuracy and recall metrix for a whole dataset(averaging through the classes)}
\usage{
compute.metrixes(t)
}
\arguments{
\item{t}{matrix - confusion matrix returned by table function}
}
\value{
list containing precision, recall and accuracy for the whole dataset
}
\description{
Function compute.metrixes given confusion matrix, computes precision, accuracy and recall metrix for a whole dataset(averaging through the classes)
}

