% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_frame.R
\name{create.articles.data.frame}
\alias{create.articles.data.frame}
\title{Function create.articles.data.frame creates dataframe from preprocesses files lokated in one directory}
\usage{
create.articles.data.frame(articles.directory, output.df.filename)
}
\arguments{
\item{input.path}{string directory with preprocessed files}

\item{output.path}{string parametr is the name (location+name) where the dataframe should be saved}
}
\description{
Files should have flat structure and should have format returned by preprocess function from the package
}
\examples{
create.articles.data.frame("C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\prepr\\\\", "C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\data\\\\")

}

