% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_df_categories.R
\name{data.compact}
\alias{data.compact}
\title{Function data.compact removes columns with 0 value (eg. is used in knn algorithm)}
\usage{
data.compact(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame with removed columns
}
\description{
Function data.compact removes columns with 0 value (eg. is used in knn algorithm)
}
\examples{
data.compact(data)

}

