% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{fit.knn}
\alias{fit.knn}
\title{Function fit.knn implements model building algorithm of k nearest neighbours}
\usage{
fit.knn(data, fact, k = 3, metric = metric.cos)
}
\arguments{
\item{data}{data.frame contains values of our training set (bag-of-words format of text)}

\item{fact}{factor containing proper classification of the training samples in data set (one per observation)}

\item{k}{number of the nearest neighbours taken into acount}

\item{metric}{function implementing the metric of the similarity}
}
\value{
object of the class ptmKnn to use in predict function
}
\description{
This fit function just remembers the train data in the model
}
\examples{
fit.knn(train.data, label, 5, metric.cos)

}

