% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_files.R
\name{flatten.files.structure}
\alias{flatten.files.structure}
\title{Loads abstracts data into another directory, flattening the structure of the files for futher processing
Data can be retrieved from http://archive.ics.uci.edu/ml/datasets/NSF+Research+Award+Abstracts+1990-2003.
Folder paths of the arguments must end with / (\\)}
\usage{
flatten.files.structure(input.folders.path, flattened.files.folder.path)
}
\arguments{
\item{input.folders.path}{string directory with inial data (ex. unzipped Part1.zip file)}

\item{flattened.files.folder.path}{string folder where flattened files structure data will be stored}
}
\value{
nothing
}
\description{
Loads abstracts data into another directory, flattening the structure of the files for futher processing
Data can be retrieved from http://archive.ics.uci.edu/ml/datasets/NSF+Research+Award+Abstracts+1990-2003.
Folder paths of the arguments must end with / (\\)
}
\examples{
flatten.files.structure("C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\Part1\\\\", "C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\Out_one\\\\")
}

