% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_df_categories.R
\name{get.df.column.indexes}
\alias{get.df.column.indexes}
\title{cs may be 'c' for categories, 'w' for words}
\usage{
get.df.column.indexes(df, cs)
}
\description{
cs may be 'c' for categories, 'w' for words
}

