% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{metric.cos}
\alias{metric.cos}
\title{Function metric.euclid calculates cosine distance between two vectors}
\usage{
metric.cos(v1, v2)
}
\arguments{
\item{v1}{vector first}

\item{v2}{vector second}
}
\value{
distance between two vectors
}
\description{
Function metric.euclid calculates cosine distance between two vectors
}
\examples{
metric.cos(c(1,2,3), c(2,3,4))

}

