% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{predict.mcNb}
\alias{predict.mcNb}
\title{Function predict.mcNb extends generic function predict for Naive Bayes model}
\usage{
\method{predict}{mcNb}(model, data)
}
\arguments{
\item{model}{list containing two fields $prior and $condprob which are vector and matrix}

\item{data}{data.frame contains values of our testing set (identical structure that was use in training set for building model)}
}
\value{
result matrix where for each sample from data set probabilities of each class are given
}
\description{
Function predict.mcNb extends generic function predict for Naive Bayes model
}
\examples{
predict.mcNb(model, data)

}

