% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_multinomial.R
\name{predict.mcTmnb}
\alias{predict.mcTmnb}
\title{Function predict.mcTmnb extends generic function predict for Naive Bayes bernoulli model}
\usage{
\method{predict}{mcTmnb}(model, data)
}
\arguments{
\item{model}{list containing two fields $prior and $condprob which are vector and matrix}

\item{data}{data.frame contains values of our testing set (identical structure that was use in training set for building model)}
}
\value{
result matrix where for each sample from data set probabilities of each class are given
}
\description{
Function predict.mcTmnb extends generic function predict for Naive Bayes bernoulli model
}
\examples{
predict.mcTmnb(model, data)

}

