% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_multinomial.R
\name{predict.mcTmnb.v2}
\alias{predict.mcTmnb.v2}
\title{Function predict.mcTmnb.v2 extends generic function predict for Naive Bayes bernoulli model}
\usage{
\method{predict}{mcTmnb.v2}(model, data)
}
\arguments{
\item{model}{list containing two fields $prior and $condprob which are vector and matrix}

\item{data}{data.frame contains values of our testing set (identical structure that was use in training set for building model)}
}
\value{
result matrix where for each sample from data set probabilities of each class are given
}
\description{
function is different approach of predicting from the same model
}
\examples{
predict.mcTmnb(model, data)

}

