% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{predict.ptmKnn}
\alias{predict.ptmKnn}
\title{Function predict.ptmKnn extends generic function predict for Naive Bayes bernoulli model}
\usage{
\method{predict}{ptmKnn}(model, data)
}
\arguments{
\item{model}{list should contain $vectors $fact, $k, $metric}

\item{data}{data.frame contains values of our testing set (identical structure that was use in training set for building model)}
}
\value{
result vector where for each sample from data set assigned category is returned
}
\description{
Function predict.ptmKnn extends generic function predict for Naive Bayes bernoulli model
}
\examples{
predict(model, data)

}

