% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_files.R
\name{preprocess}
\alias{preprocess}
\title{Function for initial preprocessing of the files
After calling the function output files will contain categories in the first line
and then cleaned body text separated by one empty line}
\usage{
preprocess(input.path, output.path)
}
\arguments{
\item{input.path}{string directory with inial data (files without any folder structure)}

\item{output.path}{string folder where preprocessed files will be stored}
}
\value{
output.path string
}
\description{
Function for initial preprocessing of the files
After calling the function output files will contain categories in the first line
and then cleaned body text separated by one empty line
}
\examples{
preprocess("C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\Part1\\\\", "C:\\\\Users\\\\vostruk001\\\\Projects\\\\Studia\\\\MOW\\\\Projekt\\\\MOW_PROJEKT\\\\Out_one\\\\")
}

