% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_df_categories.R
\name{split.categories}
\alias{split.categories}
\title{Function split.categories changes dataframe from the one with many categories for one sample to many samples with one category}
\usage{
\method{split}{categories}(df, min.n = 500, invalid.cats.hash = emptyenv(),
  min.w = 2)
}
\arguments{
\item{df}{data.frame input data frame with values and categories}

\item{min.n}{int minimalna liczba wystapien danej kategorii}

\item{invalid.cats.hash}{hashtable nazwy kategori ktore powinny zostac odrzucone}

\item{min.w}{int minimalna liczba wystapien danego slowa}
}
\value{
splitted_vectors list list that contains $data - output data $fact - output factor corresponding to the data
}
\description{
Such shape of data is required by naiveBayes function and generally recognized. Additionaly some categories
defined by used can be dropped from the data frame by passing a list invalid.cats.hash
}
\examples{
split.categories(mydataframe, 100, create.hash(c("c.OtherApplicationsNEC", "c.OtherSciencesNEC")), 2)

}

