% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_df_words.R
\name{transform.df.binary}
\alias{transform.df.binary}
\title{Function transform.df.binary trainsforms bag-of-words values (count of word in the document) to binary values (present or not)}
\usage{
\method{transform}{df.binary}(df)
}
\arguments{
\item{data}{data.frame contains values of our training set}
}
\value{
data.frame with 0-1 values
}
\description{
Function transform.df.binary trainsforms bag-of-words values (count of word in the document) to binary values (present or not)
}
\examples{
transform.df.binary(df)

}

